requirements.txt
cupy-cuda12x==13.4.1
fastrlock==0.8.3
imageio==2.37.0
lazy_loader==0.4
networkx==3.5
numpy==2.2.6
packaging==25.0
pillow==11.2.1
scikit-image==0.25.2
scipy==1.16.0
tifffile==2025.6.11
